*: use joint consensus (#386)
This PR updates raft to utilize joint consensus. Several outdated methods from ProgressTracker and benches are removed.

Most codes are ported with the same name and same location. Except switchToConfig is ported as post_conf_change as we mutates progress inline.

Tests are updated to get it passed. But more tests need to be ported later.

Signed-off-by: Jay Lee <BusyJayLee@gmail.com>